api/requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/LLM Integration
anthropic>=0.40.0

# Vector Search & Embeddings
sentence-transformers==2.7.0
huggingface_hub==0.20.3
faiss-cpu==1.7.4
numpy==1.24.3

# Database (sqlite3 is built into Python)
sqlalchemy==2.0.23

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
loguru==0.7.2